pyproject.toml
[project]
name = "authflow-cli"
version = "0.1.0"
description = "A professional OAuth2-based financial data CLI and API service."
readme = "README.md"
authors = [
    { name = "AnamolZ", email = "anmoldkl971@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "bcrypt<4.0.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
    "yfinance>=1.0",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
